Use remix-dev directly in create-remix
No need to use a shell to execute remix-dev
