+++ packages/create-remix/package.json
@@ -16,9 +16,6 @@
     "create-remix": "cli.js"
   },
   "dependencies": {
-    "shell-escape": "^0.2.0"
-  },
-  "devDependencies": {
-    "@types/shell-escape": "^0.2.0"
+    "@remix-run/dev": "1.3.1"
   }
 }
