+++ packages/create-remix/tsconfig.json
@@ -1,8 +1,9 @@
 {
-  "exclude": ["__tests__", "templates"],
   "compilerOptions": {
     "lib": ["ES2019"],
     "target": "ES2019",
+    "module": "CommonJS",
+    "skipLibCheck": true,
 
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
