+++ packages/remix-dev/cli/commands.ts
@@ -15,10 +15,10 @@ import * as compiler from "../compiler";
 import type { RemixConfig } from "../config";
 import { readConfig } from "../config";
 import { formatRoutes, RoutesFormat, isRoutesFormat } from "../config/format";
-import { createApp } from "../create";
+import { createApp } from "./create";
 import { loadEnv } from "../env";
-import { setupRemix, isSetupPlatform, SetupPlatform } from "../setup";
-import { log } from "../log";
+import { log } from "./logging";
+import { setupRemix, isSetupPlatform, SetupPlatform } from "./setup";
 
 export async function create({
   appTemplate,
@@ -75,8 +75,8 @@ export async function create({
   }
 }
 
-export async function init(remixRoot: string) {
-  let initScriptDir = path.join(remixRoot, "remix.init");
+export async function init(projectDir: string) {
+  let initScriptDir = path.join(projectDir, "remix.init");
   let initScript = path.resolve(initScriptDir, "index.js");
 
   if (await fse.pathExists(initScript)) {
@@ -84,7 +84,7 @@ export async function init(remixRoot: string) {
     execSync("npm install", { stdio: "ignore", cwd: initScriptDir });
     let initFn = require(initScript);
     try {
-      await initFn({ rootDirectory: remixRoot });
+      await initFn({ rootDirectory: projectDir });
     } catch (error) {
       console.error(`🚨 Oops, remix.init failed`);
       throw error;
