+++ packages/remix-dev/cli/index.ts
@@ -1,0 +1,2 @@
+export * as commands from "./commands";
+export { run } from "./run";
