+++ packages/remix-dev/index.ts
@@ -2,4 +2,7 @@ import "./modules";
 
 export type { AppConfig } from "./config";
 
-export { createApp } from "./create";
+export * as compiler from "./compiler";
+
+export * as cli from "./cli/index";
+export { createApp } from "./cli/create";
